Fix circular import
I am aware that this is reverting your change, Will.

I got the impression that your motivation was that
the reducer had kind of well-defined functionality and it didn't make
sense to have code related to state in this file given the name.
I don't like the circular imports though: I prefer a piece of software to be "stacked" into layers.

The fix I'm going to go for (in the next commit) is to rename
csReducer to csState. csState will contain all code to do with
with the state of the system, while store with contain code to do
with the redux store.