Applied no-cycle rule and moved eslint config
Actually found some circular dependencies in the project already! Will try to clean these up and therefore demonstrate that the rule is being applied properly.

Moved eslint config from package.json to its own config file which is a bit neater and probably more readable in the future when futher configuration is required.