Uses value rather than selected attribute
Follows advice here: https://reactjs.org/docs/forms.html#the-select-tag

Prevents displaying of value which is not actually the current value in the PV.
If there is a delay in value change propagation, this will be reflected in the value of the MenuButton, which will not change until the PV itself provides a new value.

This behaviour may need to be tweaked in the future, perhaps some note to the user that the new PV value has been sent, but seems better than just appearing to have been sent and processed.