Demonstrated extra properties dictated by the base widget
In this case, precision is passed through as a parameter for ReadbackWidgets and this is received properly by the Widget transformers, shown in the last readback reading from sim://sine with 3 digit precision.

Also fixed an off by one error in widget.

Quite happy with this so far.

Would be useful to further define the essential properties for wrappers as they seem to be fairly similar. Could then generalise the wrapper logic (would be best to define the if... logic elsewhere.

Should create some tests for this.