Significant change to widget interface
Widgets are now instantiated WidgetComponents which have been wrapped in one or more HOCs.

Widgets have been wrapped only in a macroWrapper and are not allowed to request wrapper components such as the alarmborder or copywrapper.

PVWidgets are wrapped in a macroWrapper and connectionWrapper and may have wrappers.

Widgets and PVWidgets also have their own interface which may be extended by specialised widgets where appropriate.

This change has been made and has propogated through the other files.

The next step is to remove now redundant code from the existing widgets.

Following that, work will be done to harden the interfaces using prop-types for type checking at runtime. It is hoped that the existing separation of interfaces will make this a simple change, at least conceptually.