Connection forwarder -- Register pvname's with different connections
This is quite difficult to test because the code is fairly gluey.
It is fairly transparent when this code isn't working from the
front page.

At the moment, the page will error out if e.g. the Coniql connection
hasn't been setup in `App.tsx` and the user still has to manually
set up connections in `App.tsx`.

We might to set up a configuration file at some point.

I imagine other people might have arguments for catching
errors and only display them next to the effected widgets.