Demonstrated using proptypes for run time type checking.
Front page and Readbacks JSON files have an error in them.
This error is located at runtime by WidgetToDescription which throws an error, which is then converted by fromJSON to being an error component. This is purely for demonstration purposes.
Feedback about the error is also given on the console.

Readback and Label have been converted to using the prop types, although they also have typescript information in them.

Note that typescript info can still be used for the underlying components as this will be tested during compilation, but the widgets themselves CAN eventually rely purely on prop-types.

This has not been fully implemented so far so that we can have a look at what this will actually look like. Hence the macro page having some things removed and the interface still hanging around in the Label and Readback widgets.