Runtime type checking improved
Now catches extra keys in embedded shapes by using "exact" instead of "shape" for runtime type checking.

The cost of this can be seen in macroPage.tsx, where a bunch of values have to be initialized even though they are not required. This is due to inaccuracies in the InferProps method which does not exactly characterise the behaviour of PropTypes "exact" and "shape".