Widgets now use prop types for their own special props and inherit other things when necessary
Note that the change to utils is probably not the last which will be required, as prop-types requirable returns <T | null | undefined> whereas key?: T returns T | undefined.

Have looked into changing this but I think quite a few people who are concerned with prop-types would rather not have it changed as they rely on that behaviour when building up further validators etc of their own.