Removed rule hook from Widget
This has a clear effect on performance, there appears to be something about this hook in particular which is detrimental to performance and causing re renders of Widget on every cycle.
When it is removed, this effect is not seen and, as desired, neither the Widget nor DisplayComponent are rendering.