Extract known properties and group them according to the objects **Widget** expects
Also provided a function which removes *falsey* values from the object before passing it on.
Widget styling is now provided directly.

This method also works with the runtime type checking we have been working on.

Note that because of the way React squashes its props together, the existing JSON format now works alongside this one and in principle they can probably be mixed and matched.