Fixed rendered on unmounted component
Seems that this issue was actually caused byt eh updated macros continuously causing renders when not necessary and probably triggering unnecessary fetches. Have changed this logic now so that only a change in the file location will trigger a fetch. useEffect may still be necessary in the future if the fetch is expected to possibly take a long time.