Attempt to simplify Widget.
A widget is now rendered into one div, whose location is determined
by the position prop.

Context menu and tooltip are handled in the Widget and
ConnectingComponent components respectively and no longer
need to add their own divs.

Outlines are now handled in Widget and this means that the
useCommonCss() hook no longer needs to be a hook.